Update Cms.php
In case of a page identifier is assigned to multiple stores and exist multiple times, an issue happen `A page URL key for specified store already exists.`. The reason is because the store_id is missing into the CMS Page object and the resource method _getLoadSelect is skipped. The filtering is not enough sufficient and the object is instanciated not with appropriate data

```
{
    {
      "identifier": "home",
      "store_id": "magicshop",
      "is_active": 1,
      "content_file": "home_de.html",
      "stores": [
        "magicshop",
        "magicshop_eu"
      ]
    }
}
```